Better tests and project structure
