+++ Tests/TidyJSON/Info.plist
