--- Tests/main.swift
@@ -1,4 +1,0 @@
-import XCTest
-import TidyJSON
-import Foundation
-XCTMain([ValueTests(), ParserTests(), DumpTests(), ModifyTests()])
